+++ core/src/test/java/io/questdb/test/cairo/FullFwdDataFrameCursorTest.java
@@ -1153,7 +1153,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 boolean invoked = false;
 
                 @Override
-                public boolean closeRemove(int fd, LPSZ name) {
+                public boolean closeRemove(long fd, LPSZ name) {
                     if (Utf8s.endsWithAscii(name, ".lock")) {
                         invoked = true;
                         super.close(fd);
@@ -1248,7 +1248,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 private int mapCount = 0;
 
                 @Override
-                public long mmap(int fd, long len, long offset, int flags, int memoryTag) {
+                public long mmap(long fd, long len, long offset, int flags, int memoryTag) {
                     // mess with the target FD
                     if (fd == this.fd) {
                         if (mapCount == 1) {
@@ -1260,7 +1260,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 }
 
                 @Override
-                public long mremap(int fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
+                public long mremap(long fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
                     if (fd == this.fd) {
                         if (mapCount == 1) {
                             return -1;
@@ -1271,7 +1271,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 }
 
                 @Override
-                public int openRW(LPSZ name, long opts) {
+                public long openRW(LPSZ name, long opts) {
                     if (Utf8s.endsWithAscii(name, fileUnderAttack)) {
                         this.fd = super.openRW(name, opts);
                         return this.fd;
@@ -1410,7 +1410,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 }
 
                 @Override
-                public long mmap(int fd, long len, long offset, int flags, int memoryTag) {
+                public long mmap(long fd, long len, long offset, int flags, int memoryTag) {
                     // mess with the target FD
                     if (fd == this.fd) {
                         return -1;
@@ -1419,7 +1419,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 }
 
                 @Override
-                public int openRW(LPSZ name, long opts) {
+                public long openRW(LPSZ name, long opts) {
                     // remember FD of the file we are targeting
                     if (Utf8s.endsWithAscii(name, fileUnderAttack)) {
                         return fd = super.openRW(name, opts);
@@ -1581,7 +1581,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 private int mapCount = 0;
 
                 @Override
-                public long mmap(int fd, long len, long offset, int flags, int memoryTag) {
+                public long mmap(long fd, long len, long offset, int flags, int memoryTag) {
                     // mess with the target FD
                     if (fd == this.fd) {
                         if (mapCount == 1) {
@@ -1593,7 +1593,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 }
 
                 @Override
-                public long mremap(int fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
+                public long mremap(long fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
                     // mess with the target FD
                     if (fd == this.fd) {
                         if (mapCount == 1) {
@@ -1605,7 +1605,7 @@ public class FullFwdDataFrameCursorTest extends AbstractCairoTest {
                 }
 
                 @Override
-                public int openRW(LPSZ name, long opts) {
+                public long openRW(LPSZ name, long opts) {
                     // remember FD of the file we are targeting
                     if (Utf8s.endsWithAscii(name, fileUnderAttack)) {
                         return fd = super.openRW(name, opts);
