+++ core/src/test/java/io/questdb/test/cairo/RecoverVarIndexTest.java
@@ -294,7 +294,7 @@ public class RecoverVarIndexTest extends AbstractCairoTest {
             AtomicInteger count = new AtomicInteger();
             ff = new TestFilesFacadeImpl() {
                 @Override
-                public int openRW(LPSZ name, long opts) {
+                public long openRW(LPSZ name, long opts) {
                     if (Utf8s.containsAscii(name, "str2.i") && count.incrementAndGet() == 14) {
                         return -1;
                     }
