+++ core/src/test/java/io/questdb/test/cairo/TableReaderTest.java
@@ -1859,7 +1859,7 @@ public class TableReaderTest extends AbstractCairoTest {
                 temp.of(engine.getConfiguration().getRoot()).concat("dummy_non_existing_path");
                 ff = new TestFilesFacadeImpl() {
                     @Override
-                    public int openRO(LPSZ name) {
+                    public long openRO(LPSZ name) {
                         if (Utf8s.endsWithAscii(name, TableUtils.META_FILE_NAME) && openCount.decrementAndGet() < 0) {
                             return TestFilesFacadeImpl.INSTANCE.openRO(temp.$());
                         }
@@ -1897,7 +1897,7 @@ public class TableReaderTest extends AbstractCairoTest {
                 ff = new TestFilesFacadeImpl() {
 
                     @Override
-                    public long length(int fd) {
+                    public long length(long fd) {
                         if (fd == this.fd) {
                             return Files.length(temp.$());
                         }
@@ -1913,7 +1913,7 @@ public class TableReaderTest extends AbstractCairoTest {
                     }
 
                     @Override
-                    public int openRO(LPSZ name) {
+                    public long openRO(LPSZ name) {
                         if (Utf8s.endsWithAscii(name, TableUtils.META_FILE_NAME) && openCount.decrementAndGet() < 0) {
                             return this.fd = TestFilesFacadeImpl.INSTANCE.openRO(name);
                         }
