+++ core/src/test/java/io/questdb/test/cairo/vm/ContiguousMemoryMTest.java
@@ -156,7 +156,7 @@ public class ContiguousMemoryMTest extends AbstractCairoTest {
 
             FilesFacade ff = new TestFilesFacadeImpl() {
                 @Override
-                public long length(int fd) {
+                public long length(long fd) {
                     long r = Files.length(fd);
                     if (r < 0 || trigger.get()) {
                         throw CairoException.critical(Os.errno()).put("Checking file size failed");
@@ -719,7 +719,7 @@ public class ContiguousMemoryMTest extends AbstractCairoTest {
             boolean failTruncate = false;
 
             @Override
-            public long mremap(int fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
+            public long mremap(long fd, long addr, long previousSize, long newSize, long offset, int mode, int memoryTag) {
                 if (--counter < 0) {
                     failTruncate = true;
                     return -1;
@@ -728,7 +728,7 @@ public class ContiguousMemoryMTest extends AbstractCairoTest {
             }
 
             @Override
-            public boolean truncate(int fd, long size) {
+            public boolean truncate(long fd, long size) {
                 if (failTruncate) {
                     return false;
                 }
